requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pillow==11.0.0
piexif==1.3.3
jinja2==3.1.4
sqlalchemy==2.0.36
itsdangerous==2.2.0
